change prompt when reply

+++ models/msg.go
@@ -2,8 +2,8 @@ package models
 
 type Message struct {
 	ID          uint   `gorm:"primaryKey" json:"id"`
-	SenderID    uint   `gorm:"not null" json:"sender_id"`
-	ReceiverID  uint   `gorm:"not null" json:"receiver_id"`
+	SenderID    int64  `gorm:"not null" json:"sender_id"`
+	ReceiverID  int64  `gorm:"not null" json:"receiver_id"`
 	ChatID      string `gorm:"not null" json:"chat_id"`
 	Content     string `gorm:"type:varchar(2000);not null" json:"content"`
 	MsgType     uint8  `gorm:"not null;default:1" json:"msg_type"` // 1=text, 2=image, etc.
